refactor: 경로와 요금 관련 도메인 객체 별도의 패키지로 분리
