+++ src/main/java/wooteco/subway/domain/fare/Fare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.fare;
 
 public class Fare {
