+++ src/main/java/wooteco/subway/domain/path/PathCalculator.java
@@ -1,9 +1,10 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 import java.util.List;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+import wooteco.subway.domain.Section;
 
 public class PathCalculator {
