+++ src/main/java/wooteco/subway/service/PathService.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Discounter;
-import wooteco.subway.domain.Fare;
+import wooteco.subway.domain.fare.Discounter;
+import wooteco.subway.domain.fare.Fare;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Path;
-import wooteco.subway.domain.PathCalculator;
+import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.path.PathCalculator;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.service.dto.request.PathRequest;
