+++ src/test/java/wooteco/subway/domain/DiscounterTest.java
@@ -3,6 +3,7 @@ package wooteco.subway.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.domain.fare.Discounter;
 
 import static org.assertj.core.api.Assertions.assertThat;
