+++ src/test/java/wooteco/subway/domain/PathCalculatorTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.path.PathCalculator;
 
 @DisplayName("PathCalculator 는")
 class PathCalculatorTest {
